requirements.txt
# Core numerical / ML stack
numpy>=1.24
pandas>=2.1
scikit-learn>=1.3
xgboost>=2.0
PyYAML>=6.0

# PyTorch + HuggingFace ecosystem
torch>=2.1
transformers>=4.38
accelerate>=0.24
sentencepiece>=0.1

# Experiment tracking & utilities
wandb>=0.16
tqdm>=4.66
requests>=2.31
openai>=1.0
